It did work as the command is designed to work
Python test file was located in D:\DataScience-Python3, but when I open up command prompt, it's in the default C drive. So, I was trying to use cd command to change working directory to D drive. However, cd command just couldn't work, it ended up in the C drive. It was a little bit frustrating. Here comes the magic command "/d", which enabling change to a particular directory on a different drive. Afterwards, only need to type c: or d: to change to the desired drive.